add IndicatorLEDs::restoreBlinkingLEDs()

+++ src/IndicatorLEDs.h
@@ -35,7 +35,7 @@ class IndicatorLEDs
         static const uint8_t NUM_BLINKS;        //number of blinks to stop
         static const uint8_t SCANS_PER_BLINK;   //smaller number makes blinking faster
         static const uint8_t SCANS_BLINK_ON;    //number of scans per blink that LED is on
-        uint8_t scansSinceFirstBlink;           //magic number 0 means not blinking
+        uint16_t scansSinceFirstBlink;          //magic number 0 means not blinking
     public:
         IndicatorLEDs(LED*const ptrsLEDs_L[], LED*const ptrsLEDs_R[],
                 const uint8_t TEN_KEY_OFF, const uint8_t TEN_KEY_ON, const uint8_t MF)
